Add support for a better #stink_score
This is the initial implementation and tests for the stink_score method and reporter.

I tried to explain how I calculated the `stink_score` by combining three values:

- cost
- churn
- code coverage

This is a decent way to prioritize files to be refactored in a legacy application.

There is no theory behind this formula, so please be careful. The goal of this code is to prioritize files which accumulate tech debt.