Fixes #6 by calculating max table width
The problem with the table width is related to pathnames in someone's application being "too long"
for the hardcoded value of 400. Some files are deeply nested and we are displaying all of them
in the table.

This solution finds the longest pathname and it uses that length to calculate the total width of
the table.

One side effect: You might end up with quite a wide table if you have files that are deeply nested
and/or have really really long names.