Move Console Report out of Status Reporter (#128)
* Add analysis class

* Move from class to module

Refactor Skunk analysis integration by removing the Analysis class and adding its methods directly to RubyCritic::AnalysedModulesCollection. Update related files to utilize the new methods for Skunk score calculations and reporting.

* Move Console Report out of Status Reporter

The status reporter has a different purpuse in the RubyCritic gem, they also have the specific usage of the Console Report and now we are following the same patters that gem created.

- Added support for a new console output format in the FormatValidator module.
- Refactored the StatusReporter to provide a simplified status message upon completion of analysis.
- Introduced ConsoleReport and Simple classes for generating detailed console reports of analysed modules.
- Updated tests to validate new console reporting functionality and ensure expected outputs.
- Adjusted existing tests to reflect changes in the status reporting method.

* Update StatusReporter class

This class used to return the Console message but now it does not have
that responsibility and will only return an empty string until we get
the GPA Score on Skunk, we will print it as RubyCritic does

* Update Status Sharer to use skunk analysis from the collection module

* Move Console report

* Update CHANGELOG.md